# Fix #3130: Crash at fuzzing target replacer