This fixes the specific issue flagged by the fuzzer.

It does beg the question what the expected output should be for this input, but I went with this:

```plain
{
	input:  `{}{}{}{\\\\}\\\\`,
	expect: `{\\\}\\\\`,
},
```

My reasoning is that the first three braces are valid, but match no keys so get removed. The second brace is opened by never closed and so it is not replaced. The closing brace is escaped and so we loose the escape character but retain the brace.

I've made an assumption that handling of escaped escape characters is an edge case and one that will need to be handled if requested.

I'll post benchmarks in a comment.