@billglover 2020-03-10
Benchmarks are in the same ballpark.

**Before:**
```plain
BenchmarkReplacer/no_placeholder-8         	159171307	         7.41 ns/op	       0 B/op	       0 allocs/op
BenchmarkReplacer/placeholder-8            	21945183	        56.2 ns/op	      32 B/op	       1 allocs/op
BenchmarkReplacer/escaped_placeholder-8    	 9879830	       121 ns/op	      48 B/op	       1 allocs/op
```

**After:**
```plain
BenchmarkReplacer/no_placeholder-8         	166248120	         7.24 ns/op	       0 B/op	       0 allocs/op
BenchmarkReplacer/placeholder-8            	22090062	        52.8 ns/op	      32 B/op	       1 allocs/op
BenchmarkReplacer/escaped_placeholder-8    	 9950439	       128 ns/op	      48 B/op	       1 allocs/op
```

Note: I've had to switch machines so would caution comparison of benchmarks with earlier PR #3121.