@billglover 2020-03-10
Whilst this works, something tells me there is a more elegant solution. I'm going to have a tinker offline, but don't have anything concrete enough to propose at this point.