@mholt 2020-03-11
@billglover While you may be right, it's also OK in cases like these to do incremental improvements. :) I think this is looking good.

FYI: The fuzzer caught another crasher:

```
[]byte{0x26, 0x00, 0x83, 0x7B, 0x84, 0x07, 0x5C, 0x7D, 0x84}
```

Fixing that one in this PR is optional, just let me know what you'd like to do! No hurry.