@billglover 2020-03-11
This one is interesting, would be good to fix it but I may need some pointers. I've added the following test:

```plain
{
	input:  string([]byte{0x26, 0x00, 0x83, 0x7B, 0x84, 0x07, 0x5C, 0x7D, 0x84}),
	expect: string([]byte{0x26, 0x00, 0x83, 0x7B, 0x84, 0x07, 0x7D, 0x84}),
},
```

Strangely this passes.

I haven't quite figured out how to run the fuzz tests to see if I can trigger this failure myself. Thoughts?