@Mohammed90 2020-03-11
Our CI has fuzzing regression as one of the runs. So once you push commits, it'll run them against your patch. Also you can add them as part of the unit test cases.