@mholt 2020-03-11
@billglover Honestly, we're still learning how to use our fuzzer... 😅 I just found out that the second crasher was found against 90f1f7bce75686d3a9864e1f36abf15930b3fc5f which isn't this PR. Since your new test passes, I bet this PR fixed both at the same time?

I say we just merge this and see if it catches any more.